rename base_entity to entity
