+++ postgres_declare/entities/cluster_entity.py
@@ -1,4 +1,4 @@
-from postgres_declare.entities.base_entity import Entity
+from postgres_declare.entities.entity import Entity
 from postgres_declare.sqlmixin import SQLMixin
