+++ postgres_declare/entities/database.py
@@ -2,8 +2,8 @@ from typing import Sequence
 
 from sqlalchemy import Engine, TextClause, create_engine, text
 
-from postgres_declare.entities.base_entity import Entity
 from postgres_declare.entities.cluster_entity import ClusterSqlEntity
+from postgres_declare.entities.entity import Entity
 from postgres_declare.entities.role import Role
