+++ postgres_declare/entities/database_entity.py
@@ -1,7 +1,7 @@
 from typing import Sequence
 
-from postgres_declare.entities.base_entity import Entity
 from postgres_declare.entities.database import Database
+from postgres_declare.entities.entity import Entity
 from postgres_declare.sqlmixin import SQLMixin
