+++ postgres_declare/entities/entity.py
