+++ postgres_declare/entities/role.py
@@ -3,8 +3,8 @@ from typing import Sequence
 
 from sqlalchemy import TextClause, text
 
-from postgres_declare.entities.base_entity import Entity
 from postgres_declare.entities.cluster_entity import ClusterSqlEntity
+from postgres_declare.entities.entity import Entity
 
 
 class Role(ClusterSqlEntity):
