+++ postgres_declare/entities/schema.py
@@ -2,9 +2,9 @@ from typing import Sequence
 
 from sqlalchemy import TextClause, text
 
-from postgres_declare.entities.base_entity import Entity
 from postgres_declare.entities.database import Database
 from postgres_declare.entities.database_entity import DatabaseSqlEntity
+from postgres_declare.entities.entity import Entity
 from postgres_declare.entities.role import Role
