+++ tests/conftest.py
@@ -4,9 +4,9 @@ import pytest
 from sqlalchemy import Engine, String, create_engine
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
 
-from postgres_declare.entities.base_entity import Entity
 from postgres_declare.entities.database import Database
 from postgres_declare.entities.database_content import DatabaseContent
+from postgres_declare.entities.entity import Entity
 from postgres_declare.entities.role import Role
 from postgres_declare.entities.schema import Schema
 from tests.helpers import YieldFixture
