+++ tests/it/test_schema.py
@@ -1,8 +1,8 @@
 import pytest
 from sqlalchemy import Engine
 
-from postgres_declare.entities.base_entity import Entity
 from postgres_declare.entities.database import Database
+from postgres_declare.entities.entity import Entity
 from postgres_declare.entities.role import Role
 from postgres_declare.entities.schema import Schema
